Move modifiers out of Grammar
